pipeline/requirements.txt
mariadb==1.1.10
requests==2.31.0
python-dotenv==1.0.1
